+++ js/jquery.terminal-src.js
@@ -5193,10 +5193,12 @@
                     });
                 }
             } else if (type === 'object') {
+                console.log(type);
                 finalize({
                     interpreter: make_object_interpreter(
                         user_intrp,
-                        settings.checkArity
+                        settings.checkArity,
+                        login
                     ),
                     completion: Object.keys(user_intrp)
                 });
@@ -6668,7 +6670,9 @@
                         prepare_top_interpreter(true);
                     });
                 }
-                if (get_type(user_intrp) === 'string' && login) {
+                if (is_function(login)) {
+                    self.login(login, true, overwrite_interpreter);
+                } else if (get_type(user_intrp) === 'string' && login ) {
                     self.login(
                         make_json_rpc_login(user_intrp, login),
                         true,