+++ js/jquery.terminal.js
@@ -35,7 +35,7 @@
  * emoji regex v7.0.1 by Mathias Bynens
  * MIT license
  *
- * Date: Sat, 08 Sep 2018 09:30:25 +0000
+ * Date: Sat, 08 Sep 2018 11:09:03 +0000
  */
 
 /* TODO:
@@ -3170,7 +3170,7 @@
     // -------------------------------------------------------------------------
     $.terminal = {
         version: 'DEV',
-        date: 'Sat, 08 Sep 2018 09:30:25 +0000',
+        date: 'Sat, 08 Sep 2018 11:09:03 +0000',
         // colors from http://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',
@@ -5193,10 +5193,12 @@
                     });
                 }
             } else if (type === 'object') {
+                console.log(type);
                 finalize({
                     interpreter: make_object_interpreter(
                         user_intrp,
-                        settings.checkArity
+                        settings.checkArity,
+                        login
                     ),
                     completion: Object.keys(user_intrp)
                 });
@@ -6668,7 +6670,9 @@
                         prepare_top_interpreter(true);
                     });
                 }
-                if (get_type(user_intrp) === 'string' && login) {
+                if (is_function(login)) {
+                    self.login(login, true, overwrite_interpreter);
+                } else if (get_type(user_intrp) === 'string' && login ) {
                     self.login(
                         make_json_rpc_login(user_intrp, login),
                         true,